requirements.txt
flask==3.0.2
werkzeug~=3.0.1
python-dotenv==1.0.1
click~=8.1.7
alchemical~=1.0.1
SQLAlchemy==2.0.27
pymysql~=1.1.0